requirements.txt
aws-cdk-lib>=2.237.1,<3.0.0
constructs>=10.0.0,<11.0.0
inngest